+++ app_Main.py
@@ -573,7 +573,7 @@ class App(QtCore.QObject):
             if name == 'hpgl':
                 continue
 
-            self.ui.excellon_defaults_form.excellon_opt_group.pp_excellon_name_cb.addItem(name)
+            self.ui.tools_defaults_form.tools_drill_group.pp_excellon_name_cb.addItem(name)
 
         # add ToolTips for the Preprocessor ComboBoxes in Preferences
         for it in range(self.ui.tools_defaults_form.tools_solderpaste_group.pp_combo.count()):
@@ -583,9 +583,9 @@ class App(QtCore.QObject):
             self.ui.geometry_defaults_form.geometry_opt_group.pp_geometry_name_cb.setItemData(
                 it, self.ui.geometry_defaults_form.geometry_opt_group.pp_geometry_name_cb.itemText(it),
                 QtCore.Qt.ToolTipRole)
-        for it in range(self.ui.excellon_defaults_form.excellon_opt_group.pp_excellon_name_cb.count()):
-            self.ui.excellon_defaults_form.excellon_opt_group.pp_excellon_name_cb.setItemData(
-                it, self.ui.excellon_defaults_form.excellon_opt_group.pp_excellon_name_cb.itemText(it),
+        for it in range(self.ui.tools_defaults_form.tools_drill_group.pp_excellon_name_cb.count()):
+            self.ui.tools_defaults_form.tools_drill_group.pp_excellon_name_cb.setItemData(
+                it, self.ui.tools_defaults_form.tools_drill_group.pp_excellon_name_cb.itemText(it),
                 QtCore.Qt.ToolTipRole)
 
         # ###########################################################################################################